refactor: хранение, добавление и обновление фильмов и пользователей перенесены в storage из контроллеров

+++ src/main/java/ru/yandex/practicum/filmorate/exception/IdNotFoundException.java
@@ -4,22 +4,22 @@ import lombok.Getter;
 
 @Getter
 public class IdNotFoundException extends RuntimeException {
-    private final int id;
+    private final long id;
     private final String itemType;
 
-    public IdNotFoundException(Throwable cause, int id, String itemType) {
+    public IdNotFoundException(Throwable cause, long id, String itemType) {
         super(cause);
         this.id = id;
         this.itemType = itemType;
     }
 
-    public IdNotFoundException(String message, Throwable cause, int id, String itemType) {
+    public IdNotFoundException(String message, Throwable cause, long id, String itemType) {
         super(message, cause);
         this.id = id;
         this.itemType = itemType;
     }
 
-    public IdNotFoundException(String message, int id, String itemType) {
+    public IdNotFoundException(String message, long id, String itemType) {
         super(message);
         this.id = id;
         this.itemType = itemType;
