+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -13,7 +13,7 @@ import java.time.LocalDate;
 @Data
 @AllArgsConstructor
 public class User {
-    private int id; // идентификатор
+    private long id; // идентификатор
 
     @NotNull
     @Email
